fix: rename to password
