+++ src/client/signin.ts
@@ -1,4 +1,4 @@
-import { credentialSignin, CredentialSigninPayload } from "./credentials.js"
+import { passwordSignin, PasswordSigninPayload } from "./password.js"
 import { oauth, OauthProvider, OauthSigninOptions } from "./oauth.js"
 import { init as passkeyInit } from "./passkey/index.js"
 interface BaseOptions {
@@ -10,8 +10,8 @@ export const appSignin = (options: BaseOptions) => {
     oauth: (provider: OauthProvider, oauthSigniOptions?: OauthSigninOptions) =>
       oauth(options, provider, oauthSigniOptions),
     passkey: () => passkeyInit(),
-    credentials: async (payload: CredentialSigninPayload) =>
-      await credentialSignin(options, payload),
+    password: async (payload: PasswordSigninPayload) =>
+      await passwordSignin(options, payload),
   }
 }
